+++ include/core/Window.hpp
@@ -4,6 +4,7 @@
 #include <SDL2/SDL.h>
 
 #include <string>
+#include <memory>
 
 // TODO: destroy EVERY Window member in it's destructor.
 // TODO: make THIS a Widget.
@@ -21,55 +22,58 @@
 // std::optional return for functions that may have errors.
 namespace tnt
 {
-    class SpriteComponent;
-    class Sprite;
+class SpriteComponent;
+class Sprite;
 
-    class Window final
-    {
-      public:
-        Window(std::string_view title, int xpos, int ypos, int width,
-               int height, Uint32 flags);
-        ~Window() noexcept;
+class Window final
+{
+public:
+  Window(std::string_view title, int xpos, int ypos, int width,
+         int height, Uint32 flags);
+  ~Window() noexcept;
 
-        operator SDL_Window *() noexcept;
+  operator SDL_Window *() noexcept;
 
-        SDL_Renderer *getRenderer() const noexcept;
+  SDL_Renderer *getRenderer() const noexcept;
 
-        int getDisplayIndex() const noexcept;
+  int getDisplayIndex() const noexcept;
 
-        int setDisplayMode(const SDL_DisplayMode *mode) noexcept;
-        std::pair<SDL_DisplayMode, int> getDisplayMode() const noexcept;
+  int setDisplayMode(const SDL_DisplayMode *mode) noexcept;
+  std::pair<SDL_DisplayMode, int> getDisplayMode() const noexcept;
 
-        int getWidth() noexcept;
-        int getHeight() noexcept;
+  int getWidth() noexcept;
+  int getHeight() noexcept;
 
-        Uint32 getPixelFormat() const noexcept;
-        Uint32 getID() const noexcept;
-        Uint32 getFlags() const noexcept;
+  Uint32 getPixelFormat() const noexcept;
+  Uint32 getID() const noexcept;
+  Uint32 getFlags() const noexcept;
 
-        void setTitle(char const *title) noexcept;
-        char const *getTitle() const noexcept;
+  void setTitle(char const *title) noexcept;
+  char const *getTitle() const noexcept;
 
-        void setIcon(SDL_Surface *icon) noexcept;
+  void setIcon(SDL_Surface *icon) noexcept;
 
-        int *getBordersSize() const noexcept;
+  int *getBordersSize() const noexcept;
 
-        void Render() noexcept;
-        void Draw(tnt::SpriteComponent const *obj, SDL_Rect const *srcrect,
-                  SDL_FRect const *cam, const double angle = .0,
-                  SDL_RendererFlip flip = SDL_FLIP_NONE);
-        void Draw(tnt::Sprite const *obj, SDL_Rect const *srcrect,
-                  SDL_FRect const *cam, const double angle = .0,
-                  SDL_RendererFlip flip = SDL_FLIP_NONE);
-        void Clear() noexcept;
+  void Render() noexcept;
+  void Draw(tnt::SpriteComponent const *obj, SDL_Rect const *srcrect,
+            SDL_FRect const *cam, const double angle = .0,
+            SDL_RendererFlip flip = SDL_FLIP_NONE);
+  void Draw(tnt::Sprite const *obj, SDL_Rect const *srcrect,
+            SDL_FRect const *cam, const double angle = .0,
+            SDL_RendererFlip flip = SDL_FLIP_NONE);
+  void Clear() noexcept;
 
-        void setClearColor(SDL_Color const &color) noexcept;
-        SDL_Color getClearColor() const noexcept;
+  void setClearColor(SDL_Color const &color) noexcept;
+  SDL_Color getClearColor() const noexcept;
 
-      private:
-        SDL_Renderer *renderer;
-        SDL_Window *window;
-    };
+private:
+  SDL_Renderer *renderer;
+  SDL_Window *window;
+};
+
+using window_ptr = std::shared_ptr<tnt::Window>;
+using window_ref = std::weak_ptr<tnt::Window>;
 } // namespace tnt
 
 #endif //! TNT_WINDOW_HPP