+++ include/math/Vector.hpp
@@ -21,9 +21,7 @@ namespace tnt
 struct Vector
 {
     explicit constexpr Vector(float _x = 0.f, float _y = 0.f) noexcept
-        : x{_x}, y{_y}
-    {
-    }
+        : x{_x}, y{_y} {}
 
     // template <typename T1, typename T2 = T1,
     // 		  typename = std::enable_if_t<
@@ -41,21 +39,21 @@ struct Vector
     float MagnitudeSqr() const noexcept { return x * x + y * y; }
     float Magnitude() const noexcept { return sqrtf(x * x + y * y); }
 
-    void Normalize() noexcept
+    void Normalize() noexcept(noexcept(Magnitude() > FLT_EPSILON))
     {
         float mag{Magnitude()};
         x /= mag;
         y /= mag;
     }
 
-    Vector Normalized() const
+    Vector Normalized() const noexcept(noexcept(Magnitude() > FLT_EPSILON))
     {
         float mag{Magnitude()};
         return Vector{x / mag, y / mag};
     }
 
     // operators
-    Vector operator-() const { return Vector{-x, -y}; }
+    Vector operator-() const noexcept { return Vector{-x, -y}; }
 
     Vector &operator+=(const Vector &rhs) noexcept
     {
@@ -98,14 +96,15 @@ struct Vector
     {
         return std::tie(x, y) < std::tie(rhs.x, rhs.y);
     }
+
     bool operator>(const Vector &rhs) const noexcept
     {
         return ((x > rhs.x) && (y > rhs.y));
     }
 
-    //#if __cplusplus > 201703L // operator<=> isn't available before C++20
-    //		auto operator<=>(Vector const&) const = default;
-    //#endif
+    // #if __cplusplus > 201703L // operator<=> isn't available before C++20
+    //     auto operator<=>(Vector const &) const = default;
+    // #endif
 
     friend std::ostream &operator<<(std::ostream &os, Vector const &rhs);
 
@@ -113,26 +112,26 @@ struct Vector
     float y;
 };
 
-inline Vector operator+(const Vector &lhs, const Vector &rhs)
+inline Vector operator+(const Vector &lhs, const Vector &rhs) noexcept
 {
     return Vector{lhs.x + rhs.x, lhs.y + rhs.y};
 }
-inline Vector operator-(const Vector &lhs, const Vector &rhs)
+inline Vector operator-(const Vector &lhs, const Vector &rhs) noexcept
 {
     return Vector{lhs.x - rhs.x, lhs.y - rhs.y};
 }
-inline Vector operator*(const Vector &vec, const float &num)
+inline Vector operator*(const Vector &vec, const float &num) noexcept
 {
     return Vector{vec.x * num, vec.y * num};
 }
-inline Vector operator/(const Vector &vec, const float &num)
+inline Vector operator/(const Vector &vec, const float &num) noexcept
 {
     return Vector{vec.x / num, vec.y / num};
 }
 
 inline float Magnitude(float x, float y) noexcept { return sqrtf(x * x + y * y); }
 
-inline float Dot(const Vector &lhs, const Vector &rhs)
+inline float Dot(const Vector &lhs, const Vector &rhs) noexcept(noexcept(lhs.Magnitude() * rhs.Magnitude() > FLT_EPSILON))
 {
     float angleCos =
         static_cast<float>((lhs.x * rhs.x + lhs.y * rhs.y) /
@@ -141,13 +140,13 @@ inline float Dot(const Vector &lhs, const Vector &rhs)
     return angle;
 }
 
-inline float Cross(Vector const &lhs, Vector const &rhs)
+inline float Cross(Vector const &lhs, Vector const &rhs) noexcept
 {
     return ((lhs.x * rhs.y) - (lhs.y * rhs.x));
 }
 
 // TODO: make this available inside the struct.
-inline Vector RotateVector(Vector &vec, float angle)
+inline Vector RotateVector(Vector const &vec, float angle) noexcept
 {
     float radAngle = static_cast<float>(DegreeToRadian(angle));
     return Vector{
@@ -156,29 +155,30 @@ inline Vector RotateVector(Vector &vec, float angle)
                            vec.y * cosf(radAngle))};
 }
 
-inline Vector nlerp(Vector left, Vector right, float pct)
+inline Vector nlerp(Vector const &left, Vector const &right, float pct) noexcept
 {
     return lerp<Vector>(left, right, pct).Normalized();
 }
 
 inline bool operator<=(const Vector &lhs, const Vector &rhs) noexcept
 {
-    return ((lhs.x <= rhs.x) && (lhs.y <= rhs.y));
-}
-inline bool operator>=(const Vector &lhs, const Vector &rhs) noexcept
-{
-    return ((lhs.x >= rhs.x) && (lhs.y >= rhs.y));
-}
-inline bool operator==(const Vector &lhs, const Vector &rhs) noexcept
-{
-    return ((fabs(lhs.x - rhs.x) < 1e-11) && (fabs(lhs.y - rhs.y) < 1e-11));
+    return (((lhs.x - rhs.x) <= FLT_EPSILON) && ((lhs.y - rhs.y) <= FLT_EPSILON));
 }
 
-#if __cplusplus <= 201703L || \
-    defined(TNT_ALLOW_DEPRECATIONS) // operator!= is deprecated in C++20
+inline bool operator>=(const Vector &lhs, const Vector &rhs) noexcept
+{
+    return (((lhs.x - rhs.x) >= FLT_EPSILON) && ((lhs.y - rhs.y) >= FLT_EPSILON));
+}
+
+inline bool operator==(const Vector &lhs, const Vector &rhs) noexcept
+{
+    return ((fabsf(lhs.x - rhs.x) < FLT_EPSILON) && (fabsf(lhs.y - rhs.y) < FLT_EPSILON));
+}
+
+#if __cplusplus <= 201703L || defined(TNT_ALLOW_DEPRECATIONS) // operator!= is deprecated in C++20
 inline bool operator!=(const Vector &lhs, const Vector &rhs) noexcept
 {
-    return ((lhs.x != rhs.x) || (lhs.y != rhs.y));
+    return (fabsf((lhs.x - rhs.x) > FLT_EPSILON) || (fabsf(lhs.y - rhs.y) > FLT_EPSILON));
 }
 #endif